move coordinates into separate file
